made rtmp-server usable from the root
