+++ src/rtmp-server/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  Server: require('./Server.js'),
+  state: require('./_magic.js').state,
+  config: require('./_magic.js').config,
+  partialMod: require('./utils.js').partialMod,
+};
